refactor(import): update some imports to use paths through the app
