+++ src/modules/users/useCases/CreateUser/CreateUserUseCase.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from 'tsyringe';
 import { v4 } from 'uuid';
-import { AppError } from '../../../../shared/errors/AppError';
+import { AppError } from '@shared/errors/AppError';
 import { User } from '../../entities/User';
 import { IHashProvider } from '../../providers/HashProvider/models/IHashProvider';
 import { IUsersRepository } from '../../repositories/IUsersRepository';
